src: extract several constants to constants.js
