+++ src/constants.js
@@ -1,0 +1,3 @@
+export const PIPE_ENCODE = 1;
+export const PIPE_DECODE = -1;
+export const MAX_BUFFERED_SIZE = 512 * 1024; // 512KB
