+++ src/core/middleware.js
@@ -1,5 +1,6 @@
 import EventEmitter from 'events';
 import {getPresetClassByName, IPresetStatic} from '../presets';
+import {PIPE_ENCODE} from '../constants';
 import {kebabCase} from '../utils';
 
 const staticPresetCache = new Map(/* 'ClassName': <preset> */);
@@ -24,9 +25,6 @@ function createPreset(name, params = {}) {
   }
   return preset;
 }
-
-export const PIPE_ENCODE = 1;
-export const PIPE_DECODE = -1;
 
 /**
  * abstraction of middleware
